Different learning to rank methods are now subclasses of abstract LearningToRank class
